// src/app/models/Desat.ts
import { ActeTraitement } from 'src/app/models/ActeTraitement';

export class Desat extends ActeTraitement {


         zone?: String | null;

    fi?: String | null;
   
    date_refus?: Date | null;
   
     motifReaffectation?: String | null;
   
     statut_operatonnels?: String  | null;
   
     date_reafictation?: Date | null;
   
     cog?: String | null;
   
     ui?: String | null;
      date_liv_reafictation?: Date | null;




}